+++ projects/swimlane/docspa-search/src/public-api.ts
@@ -2,6 +2,5 @@
  * Public API Surface of docspa-search
  */
 
-export * from './lib/docspa-search.service';
 export * from './lib/docspa-search.component';
 export * from './lib/docspa-search.module';