+++ projects/swimlane/docspa-core/CHANGELOG.md
@@ -2,7 +2,7 @@ CHANGELOG
 =========
 
 ## HEAD (Unreleased)
-_(none)_
+* Fix trailing slash issue in docspa-search
 
 --------------------